# Use faster retrieval of remote information for GitHub remotes