--- a/NEWS.md
+++ b/NEWS.md
@@ -1,3 +1,10 @@
+# Breaking Changes
+
+* `install_github()`'s previously deprecated `username` argument has been
+  removed. (#142)
+
+# Development
+
 * `remote_package_name.github_remote` and `remote_sha.github_remote` will
   lookup package name locally if possible and use faster REST APIs if the curl
   package is installed. This makes the remote lookup about 10x faster when the