--- a/R/github.R
+++ b/R/github.R
@@ -10,23 +10,36 @@ github_GET <- function(path, ..., host = "api.github.com", pat = github_pat()) {
 }
 
 github_commit <- function(username, repo, ref = "master",
-  host = "api.github.com", pat = github_pat(), use_curl = is_installed("curl")) {
+  host = "api.github.com", pat = github_pat(), use_curl = is_installed("curl"), current_sha = NULL) {
 
   url <- build_url(host, "repos", username, repo, "commits", utils::URLencode(ref, reserved = TRUE))
-  tmp <- tempfile()
 
   if (isTRUE(use_curl)) {
     h <- curl::new_handle()
     headers <- c(
-      "Accept" = "application/vnd.github.VERSION.sha",
+      "Accept" = "application/vnd.github.v3.sha",
       if (!is.null(pat)) {
         c("Authorization" = paste0("token ", pat))
       }
     )
+
+    if (!is.null(current_sha)) {
+      headers <- c(headers, "If-None-Match" = paste0('"', current_sha, '"'))
+    }
     curl::handle_setheaders(h, .list = headers)
-    curl::curl_download(url, tmp, handle = h)
-    readChar(tmp, file.info(tmp)$size)
+    res <- curl::curl_fetch_memory(url, handle = h)
+    if (res$status_code == 304) {
+      return(current_sha)
+    }
+    if (res$status_code >= 300) {
+      stop("HTTP error ", res$status_code, ".", call. = FALSE)
+    }
+
+    rawToChar(res$content)
   } else {
+    tmp <- tempfile()
+    on.exit(unlink(tmp), add = TRUE)
+
     download(tmp, url, auth_token = pat)
     get_json_field(readLines(tmp, warn = FALSE), "sha")
   }
@@ -49,13 +62,34 @@ github_pat <- function(quiet = TRUE) {
   pat
 }
 
-github_DESCRIPTION <- function(username, repo, subdir = NULL, ref = "master", host = "api.github.com", ...) {
+github_DESCRIPTION <- function(username, repo, subdir = NULL, ref = "master", host = "api.github.com", ...,
+  use_curl = is_installed("curl"), pat = github_pat()) {
 
   url <- build_url(host, "repos", username, repo, "contents", paste0(subdir, "DESCRIPTION"))
   url <- paste0(url, "?ref=", utils::URLencode(ref))
 
-  tmp <- tempfile()
-  download(tmp, url, auth_token = github_pat())
+  if (isTRUE(use_curl)) {
+    h <- curl::new_handle()
+    headers <- c(
+      "Accept" = "application/vnd.github.v3.raw",
+      if (!is.null(pat)) {
+        c("Authorization" = paste0("token ", pat))
+      }
+    )
 
-  base64_decode(gsub("\\\\n", "", fromJSONFile(tmp)$content))
+    curl::handle_setheaders(h, .list = headers)
+    res <- curl::curl_fetch_memory(url, handle = h)
+    if (res$status_code >= 300) {
+      stop("HTTP error ", res$status_code, ".", call. = FALSE)
+    }
+    rawToChar(res$content)
+  } else {
+    tmp <- tempfile()
+    on.exit(unlink(tmp), add = TRUE)
+
+    tmp <- tempfile()
+    download(tmp, url, auth_token = pat)
+
+    base64_decode(gsub("\\\\n", "", fromJSONFile(tmp)$content))
+  }
 }