… man/parse-git-repo.Rd