@gaborcsardi 2018-08-20
So, the thing is, in pkgdepends  (and co.)  if the package name is not the same as the repo name, you need to include the package name in the remote name.  It looks like this: https://github.com/r-lib/pkgdepends/blob/ed31a93b514c023f80ba617dedc2672f837d8271/tests/testthat/test-parse-remotes.R#L167-L170

We could maybe just include this in remotes as well? Maybe together with your heuristic, so there  won't be too much breakage?