@jimhester 2018-08-20
Yeah that sounds good to me!