@jimhester 2018-08-23
@gaborcsardi Any additional comments on this?