@jimhester 2018-08-23
I think the worst that could happen if we assume the wrong name is the local sha would be from the wrong package, so the package would potentially be re-installed when it may not have had to be.