@gaborcsardi 2018-08-23
So this is all good now, I think, right?